Add a system test for the teacher namespace
We want to be able to write system tests that drive the browser for the
eligibility check in the teacher namespace.

I opted to setup Cuprite as the driver for the browser. It is a
pure-Ruby implementation of CDP (Chrome Developer Protocol).

This is faster and less error prone than setting up the selenium-webdriver
library.