Remove turbo-rails
We haven't discussed using turbo-rails in this project. It will require
some consideration because it doesn't appear to work out-of-the-box with
the GOVUK components.

This change shouldn't commit us to turbo-rails, so I decided to remove
the gem for now, pending a further discussion with the team as to the
merits of using turbo-rails.