Add the misconduct question
We want to ask if there is any record of misconduct during the
eligibility check.

I opted to use the form object approach we took with Find a Lost TRN.

The validations for each step are contained within a form object and
this in turn delegates the persistence to a DB backed model.

Right now, only the happy path, ie. no misconduct, is implemented.

I assumed that we can implement the ineligible path in a future PR.