Add an ineligible path to the flow
Whenever someone answers 'no' to a question, we want to inform them that
they are ineligible to apply for QTS.

The message we display is specific to each question.

I opted to store the copy for the specific ineligible reasons in the
locales file. This makes it easier to manage the copy in a single place
and also ensures that translations in the future are considered.