Add solargraph
[solargraph](https://github.com/castwide/solargraph) is a Ruby language
server protocol; it provides intellisense capabilities to [compatible
editors](https://github.com/castwide/solargraph#using-solargraph).

`solargraph` can be used without adding a `development` dependency to a
project, and it provides basic Ruby language completion for core
language features. To set it up to work with our bespoke modules, gems,
and Rails, you need to bundle it and run a command. I've documented this
in the README.