Add the 'Do you have a degree?' question
We want to extend the generated code to ask the question 'Do you have a
degree?'.

One wrinkle with how we have implemented the ineligible page is that we
don't accommodate multi-paragraph copy yet.

I opted not to address that in this PR and leave it for a future
change.