Add a basic country select question
We want to allow someone to tell us the country they are recognised as a
teacher in to be able to present the most relevant information to them
about applying for QTS.

The designs call for an accessible autocomplete with a full list of
countries.

There is a small amount of complexity in building this, so we decided to
progressively build this question.

The first step is here, present a simple select box with a limited list
of countries that we can present information on.

The intention is to expand this list and, as it grows, introduce the
autocomplete feature to ensure the increasing number of options doesn't
degrade the UX.