Add govuk-country-and-territory-autocomplete
There is an existing library that implements the country autocomplete
field that is suitable for our purposes.

[govuk-country-and-territory-autocomplete](https://github.com/alphagov/govuk-country-and-territory-autocomplete)

The country list is provided by this library. An assumption here is that
the list it provides is suitable for what we need at the moment.

With this addition, it is no longer possible to reach the ineligible
country page. We're assuming that we will address that in a future
change when we have defined what makes a country ineligible for the
purposes of this check.

I had to add a style for the list of options that renders when using
autocomplete. The default styles appear to rely on the default font
styles, which for some reason aren't being set.