Remove historic countries from autocomplete
The `govuk-country-and-territory-autocomplete` is designed to cater to
a wide variety of "Country picking needs." One example is the situation
where an old person is specifying the country they were born in. Those
countries might not exist anymore.

However, in the case of our service, those countries can be confusing.

This approach mirrors those taken by
https://github.com/DFE-Digital/npq-prototype/pull/8/commits/f1c44082606f83efbb8351fe9cac4b5a75abf5d9
and https://github.com/DFE-Digital/npq-registration/pull/308.