Fix the country autocomplete error state
When a Rails form has a validation error message, the form builder
outputs a different ID for the element.

This breaks the autocomplete JS because it is targeting the original ID.

While making this change, I took the opportunity to wrap the
autocomplete function with a check that the relevant HTML element is
present to avoid it erroring on other pages that don't need it.