Add a feature flag for :service_open
We want to control access to the service for the general public.

I have introduced feature flags, as per our approach in [Find a lost TRN](https://github.com/DFE-Digital/find-a-lost-trn/pull/57).

I chose to defer the way we manage our ENV variables in development so
that we can have a dedicated PR for that. For the time being, the
default values are test/test.

Also, the support interface to manage the feature flags will be
implemented in a feature change.