Add Country model
This model will be used to store customised information for each
country. In this case, we'll use it to determine whether the country is
a legacy country or not, but in the future we might have custom text
that appears at the end of the checker.

Eventually, I also imagine we'd build a support interface so staff can
manage these countries and their customisations.