Use Country model
This changes how the legacy countries are determined by using the new
`Country` model. I've intentionally not added a foreign key between the
`EligibilityCheck` and `Country` as we may want to change this in the
future so the foreign key is on a potential `CountryRegion` model, but
that's one to decide later.