Add useful database seed
This prepares the database in a way that reflects reality. It's not
intended to be the source of truth for the data, and I don't expect us
to update this over time, but for now it provides the checker with
enough information to develop locally with a good representation of
production.