Add a completed state for an eligibility check
We want a way of being able to tell whether someone completed an
eligibility check.

I considered storing this as multiple fields, to represent the different
states a check could be in, eg. eligible_at and ineligible_at.

This seemed like overkill and introduces slightly more complexity than
the simpler approach of storing the timestamp of when the person reached
either the eligible or ineligible page.

We can calculate the eligibility status from the existing fields, so
having multiple timestamp fields seemed redundant.

I also removed the session value representing when the flow was
completed. This was something we brought across from Find but isn't as
relevant to the check flow.

If, in the future, we do need to know if the final screen was displayed,
we can use the presence of the completed_at field to determine this.