Add a ValidationError model
Each `ValidationError` is a record of a failed validation attempt. It is
associated with the polymorphic relation `record`, in order to be
flexible enough to be used with any class.

The assumptions here are that a form object was used to define the
validations. We store the name of the form object as a string because
there is no persistent record of this. The intention is that records
will be grouped by the form_object value when displayed in the support
interface.