Refactor PerformanceStats into a separate module
Paired with @David-Feetenby.

This is the first PR of a bigger refactor of the Performance controller.

It extracts most of the logic in the controller action into a
`PerformanceStats` lib class. The class has three private methods that
each serve to calculate the "count" and "data" for each of the sections
in the performance page.

This isn't finished, but it's reasonable to review and merge it at this
stage (when the logic looks 99% the same as before), before we tack on
further improvements (in future PRs):

- [ ] Instead of a `since` parameter, we should pass in a time slice,
  like `1.week.ago.beggining_of_day..Time.zone.now`
- [ ] Following up, `until_days` could just be called
  `number_of_days_in_period` and can be calculated inside the stats
  itself (and not passed in)
- [ ] We should do `0...until_days` (notice triple `...`) to stop us
  from having to do `days - 1` which is confusing
- [ ] We should use `1.day` instead of `3600 * 24`
- [ ] Think of better instance attributes names