Add an engine to record validation errors
We have previously recorded validation errors as a concern of the
current app. The implementation was similar across projects while also
having some details specific that prevented re-use in other apps.

This change introduces a more agnostic version of validation error
logging as a Rails engine. The goal is that this is a step towards being
re-usable across multiple projects.

To that end, some assumptions are baked in:

* the concern provided by the Rails engine will be included in a form
  object that defines validations
* the form object holds a reference to a database-backed object that is
  being validated, called `record`
* every call to `valid?`, `validate` and variations of those methods
  will result in a `ValidationError` being created when errors are
  present