Send users to legacy service unless feature enabled
The `/teacher` entry point will now take users to the legacy service,
unless a feature flag is set to allow users to go to the start page.

This doesn't disable access to the non-start pages, to allow for easy
testing by going direct to `/teacher/start` but we could disable the
other pages if we felt it was necessary.

I've also decided to use a different feature name to the existing
`service_open` one as that's reserved for whether the service can be
accessed in the first place.