Add custom content fields to regions
These will be used to determine which content is shown to the user on
the eligible page.

I decided to add these fields on to the region model directly as it's
easier to get started that way, but I imagine in the future we might
want to add a `TeachingAuthority` model with these fields on where
multiple regions can link to the same teaching authority.