Render country name with 'the' prefix
Certain countries, for example 'United States' is typically written as
'the United States' in a sentence. This PR adds a hard coded list of
countries where it is necessary to add a 'the' prefix before the name.

This solution is quite a simplistic solution as it only works for
English but since our content will always be in English I think that's
okay. I did look at using a Gem like `countries`, but it doesn't provide
us with an easy way to check whether to include a 'the' or not, instead
we'd have to use its names instead of the GOV.UK autocomplete component
ones.