Add a worker CloudFoundary app
This will be used to run Sidekiq and process background jobs. I've
decided to use the same memory, disk quota and instances as the web app,
but if we decide this isn't approriate in the future we can always
change it. I've also set the deployment strategy to standard as we don't
need two Sidekiq processes running in parallel during deployment, and it
keeps things simple.