Enforce eligibility question order
Currently, we don't enforce the order of the steps in the eligibility
check and if someone happens to know a valid URL, they can enter it
and effectively skip steps in the question flow.

This determines the furthest step that has been completed in the flow
based on the presence of the attributes.

The system needs to allow people to go to a previously completed step as
well.

This is based off
https://github.com/DFE-Digital/find-a-lost-trn/pull/119.