Disable toggling :service_open on/off in prod
Now that the service is live, the button to switch `:service_open` back
off is dangerous. It can lock users out of the service, and cause an
incident.

This PR adds some conditional logic to display an alternative, disabled
button with an explanation for features that we still want to toggle in
other envs, but not in prod.