Add staff model
This adds a model which represents our staff users, those who are
assessors and have access to the support interface.

I've enabled all the relevant sounding Devise modules, I figured it was
better to start with them enabled, and we can always remove them if we
find they're not useful in the future.

I also had to add a `devise_for` in the routes to allow the model to be
created without Devise complaining, even though these controllers don't
work yet.