Add teacher model
This adds a model which represents our teacher users, those who will
apply for QTS and log in using a magic link mechanism.

Since we won't be using many of the Devise modules for this model, it's
quite a basic version for now, and we can add more fields in the future.