Add rate limiting
This adds a basic rate limit for all users with the same remote IP
address of 120 requests in a 1 minute period. I've chosen this number
fairly arbitrarily, but I reckon it's a reasonable start.

This morning we've seen large numbers of requests coming from the same
group of IP addresses, and they're making around 4-5 requests per
second, so they will be stopped by this rate limit.

I've chosen some of the most common paths we want to protect in the
tests, rather than running the tests for every possible route. We can
add to this list in the future.