Disable DfE Analytics by default in development
Most likely in a development environment we don't want DfE Analytics to
be enabled, so this should be default. Instead, users need to explicitly
turn it on if they'd like to use it.

This makes sense as it means out of the box the application runs with
reasonable defaults, rather than needing the user to create a
`.env.local` file and disable BigQuery.