Make "Start now" button a submit button
This allows us to POST to an endpoint which initially creates the
`EligibilityCheck` instance. This means the `created_at` time is a
better representation of when a user actually started using the
service. Currently, the eligibility check is only created after a user
answers the first question, which means the duration performance metrics
are slightly skewed.