Calculate duration for users who've answered all questions
At the moment we calculate the duration stats for all completed
eligibility checks, but this includes ones not in the private beta which
means users are immediately taken to the eligible or ineligible page
after choosing their country. This skews the data and suggests that it
takes no time at all to complete the journey, when in reality we only
care about users who are filling in all the questions.