Add DfE Analytics check
This ensures that we avoid making changes to the models without making
the equivelant changes in DfE Analytics.

I've added this to the RSpec job as there doesn't seem to be a nice way
of re-using services and steps in GitHub Actions without building more
custom actions, but even then it's not possible to share the services.