Remove old eligibility checks when removing a region
There's not much else we can do with these eligibility checks if we're
removing a region for the system. We think it's okay to delete them
because:

They'll most likely be non-private beta countries, so we don't care
so much about those analytics in the performance dashboard, and
we've got the data in BigQuery to see the historical data.