Refactor how eligibility is calculated
This ensures it matches the `scope`, and only shows the ineligible
reasons for questions which have been answered, allowing us to better
support questions which can be skipped.

This fixes an issue where the misconduct text wasn't showing if a user
in ineligible due to their country, and it was considered a reason for
ineligibility since the question hadn't been answered.