Be explicit about country name usage
There are some places where we want to show the country name (even if
there's a named region), and some places where we want to show the
region name. I've fixed where this is incorrect (in the completed
requirements question) and moved where we call `CountryName` to be
explicit about what we want to see in each place.