Disable completed requirements question
This temporarily removes the question as we're not sure it works in a
way that want just for the time being. It's nearly there, but it doesn't
work perfectly and we need to take it down to give it more thought.

We can reinstate once we've got the logic properly sorted.