Add an ADR about storing files
This hopefully answers all our questions about how we're going to store
files, allow users to upload files, and model these in our database. It
should give us the confidence to proceed with building the service
knowing that when we come to support file storage the questions have
already been asked and answered.