Generate application references
This generates a reference number for the application automatically when
creating a new application. The numbers are integers, but I've made the
field a string so we can change the type in the future. The main reason
we might want to do this is because a sequential reference number means
we need to lock the table each time an application is created, this
could cause performance issues down the line.