Add age range question views
This builds the views for the age range question where we ask a user to
submit the range they're able to teach.

I've built a "Check your answers" page here even though there isn't one
in the design and this part is changing, and for now it's consistent
with the other questions.