Add "Save and come back later" button to forms
This adds a new button next to the continue button on each form allowing
users to save what they've entered, and then return to the overall
application form page rather than continuing to the next question.

I've implemented this in a way that it can be re-used across all
controllers easily by creating a shared view and putting a
`redirect_to_if_save_and_continue` method in the base controller.