Build alternative name question view
This adds a new view for the alternative name, where we ask the user if
their name is different to the one on their identity documents.

I've refactored the controller as we've now got two different edit forms
within the same controller, so we need two separate methods. It feels
like there's a lot of code duplication here, but actually a lot of the
individual fields and variables are different.