Attach documents to application forms on create
This makes it easier to work with a document instance as we can be sure
that an associated document exists for each possible type, without
needing to create one when a user clicks on a link to go to a section.

Strictly speaking, the name change document doesn't need to be here
since it's optional, however treating it the same as the other documents
keeps it consistent and simpler to work with.