Add EligibilityCheck#location
This allows us to reverse the country selection form, as we go from a
fully qualified country/territory to a country code. To go back to the
country/terriority we need to look up the country code.

I've added this method so we can pre-select the country selection form
when starting an application, if there's an eligibility check in the
session.