Fix use of association in migration
We can't rely on the association here as it may have been removed from
the models when the migration runs. Instead we need to be explicit about
fetching the related model.

We don't need to use a join here as in all environments except dev,
there aren't any application forms.