Refactor how we authenticate as a test user
This partially reverts a14b7b587b47abf905f7fae6beaeb3609b3dd2f7 and
28a9936571e1f406bdd6a3a84cbfc5557952eb74 to try and solve an issue where
Devise seems to be affecting the session, which means that when the
service is not open it's not possible to complete an eligibility check.

This refactors how we authenticate as a test user to instead not use
Devise in the base application controller and stick with simple Rails
HTTP Basic authentication. Some benefits of this is that we don't need
to get Devise involved for the eligibility check, and it simplifies the
staff auth strategy. It also means that if someone were to authenticate
using the test credentials, they wouldn't get access to the support
interface (if we accidentally left the feature flag on for example).

There is a downside which is that authentication as a support user is
duplicated in two places, but I think that's an okay trade-off. However,
I do like that it restricts the staff authentication strategy to just
authenticating as a staff user, with only access to the support
interface.