Move "Add another" question to separate page
Currently at the bottom of the work history and qualification summary
page we've got a form asking the user if they'd like to add another,
with a continue button.

After design and research we'd like to change this so instead the form
appears on its own separate page after the user presses the continue
button.