Disable flash messages for teachers
Devise sends a lot of flash messages which we don't want to see as
they're not part of our designs. Unfortunately there isn't a way of
turning these off with configuration, so we have to override methods.

We can't remove the locale strings as we still use these flash messages
for the staff authentication which is closer to standard Devise.