Specify GOVUK_NOTIFY_TEMPLATE_ID_DEVISE in Terraform
This sets the environment variable via Terraform so we can customise it.
The dev and review environments are going to use a template where
personalisation is visible in the logs, whereas the other environments
will use one where personalisation is not visible. This ensures that we
don't leak any secret tokens in the logs.