Review analytics config files
I did this by regenerating the DfE Analytics config files from scratch
and then moving the entries from the block list in to the approriate
place. This ensures there are no duplicates between the block list and
send list, and I've also put them in alphabetical order.

The main difference is that we'll now be sending more data to analytics
(which will require a follow up change to the dataform repo) and we're
clearer on which fields are PII.