Refactor personal information "Check your answers"
This moves the view to its own route (`/check`) and changes the root
route to redirect the user to the check route. This will allow us in the
future to send users directly to the "Check your answers" page where
necessary and give us better control over the flow rather than
automatically redirecting.