Only show "Check your answers" if section is complete
Currently we take the user to the "Check your answers" page if they've
started a section, however this is confusing as we expect users will
want to continue filling in the form.

This has also involved changing a few of the redirections as we can now
explicitly send people to the check page where approriate, and not
otherwise.