Move FactoryBot out of test gem group
We are now using FactoryBot to populate example data in review apps (and
possibly other non-production environments). These environments are
currently running in Rails.env.production? so we need the gem in the
production gemset.

We are discussing how to better handle this but this is the simplest
thing for now to unblock review deployments.