Create example users as 'confirmed'
Confirmation emails were being sent for each user created with the
example_data:generate task. We don't want to disable the emails in
review as we may want to test that functionality so create the example
users as 'confirmed'. They would be in any case if they had started an
application.