Set correct `HostingEnvironment.host` for review
The `HostingEnvironment.host` is used to generate the login email URLs.
For all other envs the env name is the prefix and the URL is a
`education.gov.uk` one but review apps are prefixed with the application
name and have a cloud apps URL so need to be treated differently.

This will fix the login and confirmation email links.