Add ApplicationFormHelper
This adds helpers for extracting information from application forms.
Specifically I've added a method for generating the full name and for
generating rows useful in summary cards or description lists.

I've put these in helpers rather than in the components directly as I'm
anticipating that we'll need to do this sort of thing in lots of places.