Add DescriptionListComponent
This renders a defintion list using GOV.UK styles, based on the
component in "find-teacher-training":
https://github.com/DFE-Digital/find-teacher-training/blob/main/app/webpacker/styles/components/_description-list.scss

It accepts the same data structure as the summary card component so the
same data can be used in both components. I've decided _not_ to use
slots as it feels unnecessary but we should refactor this to use slots
in the future.