[802] Add ApplicationFormState
A class to handle state transations for the application form. Each state
transation is immutable preserving the state history.

State changes can be explicit e.g. the application form is submitted and
as part of that we change(to: ApplicationFormState::SUBMITTED)

or...

implicit. The review process will have a task list of review steps.
The overall state may be a composite of those step states e.g. if all
are complete then the application form might be REVIEW_COMPLETE etc.

The state is also stored on teh ApplicationForm itself as a string
otherwise filtering on state will start to be annoying.