Add filtering on application form state
This adds the ability for a user to filter the application forms on the
state, where the checkboxes include a count of each application form so
users can see how many application forms there are in each state.

I've hard coded the list of states as we don't want to show draft
application forms in the results.