Ensure clearing country box resets filter
At the moment there's a bug with the accessible autocomplete component
where clearing the box doesn't get recorded correctly and therefore the
previously selected value continues to be sent to the server in the POST
request.

To get around this, we need to include the original text in the form
data and use that to determine whether someone is trying to clear the
autocomplete component.

We also need to consider users who have JavaScript disabled and are
therefore shown just the select box, and for that we allow the original
location param to be used if the location_autocomplete param isn't
present.