Generate upload URLs which expire
This changes how we generate URLs to uploads to make sure they expire,
meaning if the URL is shared it will only last for 5 minutes.

These URLs are still public, so we might want to consider putting it
behind authentication, but that will slow down the requests as rather
than taking the user directly to Azure, we'll have to handle the request
ourselves first. We can discuss this and add that improvement in a
separate PR.