Add an ADR on assessments
This adds an ADR which described our approach to implementing the
assessments. The bulk of this ADR is pseudocode that was developed
during a spike in to how we might implement this, which may be a little
unusual for an ADR, but I think it conveys the information in a
useful way.