Allow submitting an assessment section check
This adds a form to the assessment section page allowing an assessor to
submit the details of an assessment section. It doesn't render any
failure reasons yet since those will come as we implement each section.

I haven't added any tests either as I expect we'll do this on a section
by section basis as well.