Include timeline events when submitting an application form
When a form is submitted it should include a timeline event recording
the state change from draft to submitted.

I've decided with adding this to the traits and it hasn't caused too
many issues with the tests, we should monitor this though in case it
starts to become a problem.