Add "Cyprus (Non-European Union)"
We need to add this as a separate country as teachers coming from this
area of Cyprus aren't eligible for QTS at the moment, and in the future
it'll require separate information it we do start allowing teachers from
there to receive QTS.

This has been modelled as a new country rather than as a region of the
"Cyprus" country as we need to send a unique country code for DQT for
either territory, and it allows us to make one territory ineligible and
one eligible.

There's no official country code for the non-european union part of
Cyprus so I've created one which I think makes sense.