Limit which recommendations are available
We should only be able to award an assessment if all the sections are
complete, or decline an assessment if any of the sections need an
action. This commit changes the form to only show radio buttons for the
recommendation which the user is allowed to choose, based on those
rules.