Add note model
This adds a new table and model for representing assessment notes, which
assessors can use to write free text associated with an application.

I've associated notes with the application forms directly since they can
be added at any point in the process, not necessarily by the assessor
performing the assessment.