Add create or sign in route
This adds a new route, and replaces usage of the existing route to it,
which will be used to render the page asks the user if they need to sign
in or go through the eligibility checker.

This allows us to use the existing route for a page where the question
isn't asked.