Restrict access to non-draft application forms
Once an application form has been submitted, we don't want teachers to
be able to go in and make changes to the application form. We will need
to open changes up again when we implement further information, but
until that's built we can lock the application form (and all associated
models) as soon as it's submitted.