Validate work history dates
* Start date required
* Start date must be valid
* Start date must be in the past
* End date required if not still employed
* End date must be valid
* End date must be after start date

Entering an invalid date e.g. 13 2022 into either date was raising an
error. It doesn't seem possible to validate these as strongly typed
dates as it throws a ActiveRecord::MultiparameterAssignmentErrors error
because the month is out of range. This happens before validation so it
doesn't look like we can check first.

Remove the `date` typing on the attribute and handle as a hash of
values.