Handle potential DQT network errors better
This changes how we create the DQT TRN requests in our database to
ensure there's no potential for the TRN request to exist in DQT and us
not know about it. We ensure that before making the request, a record is
created in our database so that if anything fails further along, we will
always know what request ID we used and can retry with the same one.