Refactor TRN request jobs
This refactors how we handle the TRN request jobs to better handle the
situation where a network error has occurred and we don't know the state
of the DQT TRN request. It also simplifies the code to make it easier to
read and understand, removing the need for two jobs and services and
instead consolidating the work in to a single job.