[1076] Hide 'state trained in' unless present
* Don't return it from the helper if it doesn't have a name

The application form factory was creating application forms with
nameless regions by default. This caused the index of the rows in the
summary to change and we've used that in the page object to get the
status. Making the region have a name by default seems a pragmatic way
to avoid having to deal with this in all of the tests that broke and it
doesn't break any new ones.