Update DQT retry attempts
* Only retry 3 times for Faraday::Error

While these retries are happening any error isn't reported to Sentry. 10
times backed off exponentially leaves us with too long of a delay before
we see unrecoverable errors. While we learn how the integration is going
to behave and understand enough to be more granular reduce the retries
so as we can intervene.