Convert age range and subjects notes
Rather than being a foreign key to a note, we should record these as
strings on the assessment. This means we can keep the definition of a
note as a generic piece of text that was added by an assessor, and
instead we can have a dedicated event for age range and subject.