Switch to Redis Session Store
This changes the session store to use our Redis instance rather than
relying on storing the data in a cookie. It allows us to store more data
in the session than we can in the session store and we anticipate we'll
need this to support our back links stack and filters.