Allow teachers to have many application forms
This changes the relationship between teachers and application forms
making it clear that it is possible for teachers to have multiple
application forms, which is the case when they might have been declined
and then wish to apply again.

We're not expecting there to be a need for a teacher to view their
multiple application forms so we only show them the most recently
created one, however we need to keep the old ones for assessors.