Make assessment_id column not nullable
This is a foreign key column on further information requests that link
to assessments, but it doesn't make sense to be nullable since it should
never be null in normal operation.

We're seeing issues in Sentry related to this being null in the review
environment:
- https://sentry.io/organizations/dfe-teacher-services/issues/3854623435/?project=6426061&query=is%3Aunresolved&referrer=issue-stream
- https://sentry.io/organizations/dfe-teacher-services/issues/3854775780/?project=6426061&query=is%3Aignored&referrer=issue-stream

It's not entirely clear to me why the data is in this state, but
changing the column should at least tell us when the data ends up in
this state.