Add application_forms:set_awarded_at
This is a Rake task which we can use to backfill the new awarded_at
column which was added in a512175ddfd82326f2a13db822482bcec0a6d25a.

Once the column has been deployed to all environments, this Rake task
can be run to backfill the value. I've tested the task manually in the
dev environment and it works, once this has been run I'll delete the
Rake task.