Add references to example data and personas
This adds the ability to see the reference requests on the personas page
as they'll automatically be generated as part of the example data. I've
also decided to slim down the number of different states of application
forms we generate as I'm not sure it's useful at the moment to generate
so many.