Remove retry_on calls
It looks like the jobs aren't retrying and the theory at the moment is
that Sidekiq and ActiveJob are conflicting, so we're going to take out
the `retry_on` method calls and instead fall back to the default Sidekiq
behaviour and see if we start to see jobs retrying.

I suspect this change will start reporting errors to Sentry whenever a
retry occurs, but we can filter them out if we see that.