Add eligibility_skip_questions to support interface
This adds a new checkbox to the support interface which allows
controlling whether the country skips the eligibility questions (which
works in a similar way to the "legacy" field, but still shows the
eligible content).

I'm expecting the "legacy" field to be removed when the new regulations
are in place.