Fix issue where form for further information is missing
This makes a slight change to the logic to decide whether the form on
the edit further information page for an assessor appears. This should
fix a bug where the form disappears after the pass/fail has been
selected, but the rest of the flow hasn't been followed through yet.