Add application_forms:set_declined_at
This is a Rake task which we can use to backfill the new declined_at
column.

Once the column has been deployed to all environments, this Rake task
can be run to backfill the value. I've tested the task manually in the
dev environment and it works, once this has been run I'll delete the
Rake task.