Add temporary maintenance banner
The service is going to down for some planned maintenance tomorrow so we
should add a banner warning users this is the case.

We can revert this PR once the maintenance is complete - in the future
we might want to think about building a mechanism for this in the
support console.