Redirect users to incomplete work history
Before taking them to the "add another" page. This matches the behaviour
we've got for qualifications, and fixes an issue where an incomplete
work history item can't be completed until a user selects "no" and adds
another. Once the content on the "add another" page is improved to
explain why we haven't received at least 9 months, it should be clear
why the user has ended up on this page.