Split award/decline recommendation controller
This changes how the award and decline recommendation controller works
by splitting it up in to two separate controllers. At the moment the two
journeys are very similar so it makes sense to keep them in the same
controller, but in the future they will be quite different and separating
them now should make it easier to build that.

I've also moved when the assessment recommendation is saved to the
database so the data is a better representation of when the
recommendation was made.